Significantly improved definition of minBalance
It now fits waaaay better with the rest of the development; we can state it as
a single definition, reusing everyting that came before, rather than the 8 new
definitions we needed before. Lemmas that relate this new definition to the
rest are now also much simpler. This is WIP, refactoring not totally done yet.